requirements.txt
absl-py==1.1.0
astor==0.8.1
cachetools==4.2.4
certifi==2022.6.15
charset-normalizer==2.0.12
gast==0.2.2
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.46.3
h5py==3.7.0
idna==3.3
idx2numpy==1.2.3
importlib-metadata==4.11.4
keras==2.9.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
Markdown==3.3.7
numpy==1.16.4
oauthlib==3.2.0
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
opt-einsum==3.3.0
Pillow==9.2.0
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
requests==2.28.0
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
tensorboard==2.0.2
tensorflow==2.0.0
tensorflow-estimator==2.0.1
termcolor==1.1.0
typing_extensions==4.2.0
urllib3==1.26.9
Werkzeug==2.1.2
wrapt==1.14.1
zipp==3.8.0
